Simplify input of tests for multi-bit multi-way Mux circuits
Before this commit, it shared the logic of other circuits: using a various number of inputs that are then used as a cartesian product to create all the tested cases. But for a Mux, it doesn't make any sense, as we just want a way to differentiate the inputs to see if they are correctly selected by the SEL bits.

So, this commit just create a single input for each way and then test all SELs  on these.

It reduce the work from O(m*n^2) to just O(m) without sacrificing testing of the Mux gate.