Move the first parameter of 'connect_output()' to the end.
Before this commit, it was "Circuit Output Id, Component Id, Component Output Id", so the end output was the first argument. This commit put this parameter as the last one.

It's more logic and unifies with others : the 'connect_*()' are always on the format :
	Source Component Id, Source Port Id, Target Component Id, Target Port Id
so, it allows to visualize the connections as read as left-to-right.

(As before, the 'Component Id' is not included if it refers to the circuit itself)