Add two new core "gates": the constants ONE and ZERO
This change has a lot of repercussions:
- Add these circuits construction to the base ABC builder
- Ignore them in the encoders
- Add them as base in the decoders
- Add them in the simulators
- Add them in the tests

Bonus refactoring:
- In the tests, do not use raw indices to get the correct simulators. Instead, use a mapping at the top of a file. It allows *maybe* for more flexibility, and it's easier if the circuit library changes.

I was kinda dissapointed to be forced to add two core circuits. I really wanted to really have NANDs... But let's say they're very special circuits, not really gates. The main reason is that an incrementer is so much easier to implement with a constant. Another reason is that I've seen some paper that uses logic circuit using Deep Differentiable Logic Gate Networks, and the ZERO and ONE gates are necessary.

My main fear would be that the bit-packed encoding would suffer... but with a little luck to avoid threshold effects, there's **no** difference of size! Both the current playground and nand2tetris library does not grow of a single bit!

So, I'm very happy about it!

https://google-research.github.io/self-organising-systems/difflogic-ca/