Fix and simplify the de/encoders :
Instead of always returning a bitarray, it can also returns a list[int]. That's the "native format" of the DefaultDe/Encoders. It's quite nice because it removes manipulations of ints into bytes. Also, it removes a bug when a big enough circuit (like flattened nand2tetris ALU) can't be represented as a one byte number.

This bug is kinda annoying to compare the size of the encoding. Before it was always a single byte by integer, but now it can be several bytes. I don't want to have dynamic byte counting, and I don't want to have the real Python size of ints, it's way too big to be coherent with the encoding. As an unsatisfactory middle ground, every int of the default encoding is considered to be 16 bits long. 65 536 components/inputs/outputs is more than enough (for now ?)

Also, do some light refactoring for good practices like initializing the member attributes into the __init__() and not in the main method.