WIP : Adding a flattened encoder
The principle of this flattened encoder is that a circuit composed only of NANDs (with optional ZEROs and ONEs) have a lot less moving part, so it's really easy to need less bits.

The main thing is that it's not necessary to encode the wiring output: NAND have only one output ! Same for ZEROs and ONEs.

Another thing I implemented, that *maybe* would've worked for the other encoders, is to have a progressive indexing of NANDs. As long as we have less than n NAND, we encode their index in log2(n) bits. For example:

For the first two NANDS, their index will be encoded as 0 and 1 : in 1 bit
When the third NAND is encountered, all the index will be encoded in 2 bits. The first two will be 00 and 01, and the third 10.

It's a *huge* optimization, especially for this encoding where there's *a lot* of gates.