WIP Flattened decoder (probably will delete everything :((((
Everything is a lie. I should've known 2300 nands encoding would never be able to compete with recursively defined circuits. My hubris made me change a line because I was certain that the new encoding would be better than the last one. As such, the encoding was wrong, as well as the stats.

At least, it made me implement a smart idea that can be implemented in the bitpacked encoding