Be smarter with BitPackedDe/Encoder compression regarding component's outputs bitlength
Strangely, the encoding of the component's output used the bitlength of the **circuit**'s outputs, not their own output's bitlength.

Well, there was a reason I think: the decoding is now much more brittle: the circuit definition *must* have the component in the correct order, otherwise the decoding fail. It's a bit sad, because I took care of having inputs and connections stashed to avoid this kind of problem.

But well, the compression size is the priority.

So yeah, BREAKING CHANGE.