Add a flattener for circuits, and add it to automated tests
The flattener remove the nested nature of the circuits to only keep the cores (NAND, ZERO, ONE)

This was supposed to make a really performant (size-wise) d/encoder... But it's a failure (see the branch flattened_encoding). For the HACK ALU for example, the numbers of NAND are really too much, so the number of bits length necessary explodes and the advantages of having only one type of circuit are just not enough. Maybe a Rolling/Implicit-aware
D/Encoders would help. For now, put it to rest.

Oh, but the good thing about the flattener is that it's really faster to simulate. Let's see if I integrate it into the SimulatorFast or into a new Simulator.