Loosen the D/Encoder output format list[int] is also accepted, not only bitarray
The DefaultEncoder works on integers. They're casted into a bitarray, with the precondition of them being lower than 255: one-byte long. For big circuits, it isn't enough. So, instead of doing special computation, a list[int] is returned: much simpler to take care of.

The main drawbacks is to compute the encoding size. To avoid headaches, each integers is supposed to be 2-bytes long... For now.