Optimize the BitPackedEncoder by correctly using the bitlength of component's output
Before, for the component wiring, the output index was encoded with the bitlength of... the circuit output. It's not only wasteful, but incorrect. But because the circuit outputs count was always ≥ its component outputs count, it didn't cause any bugs.

Now, the index is correctly encoded in n-bits, with n begin the bitlength (with offset) of the component's output.

As a side-effect, optimizing the circuits is necessary: this fix makes the definition of a circuit more brittle (the component must be declared in a specific order). Topological order remove this weakness.